refactor: move product data lifecycle to native core

+++ native/src/EmsScout.Domain/AirConditionerCard.cs
@@ -3,7 +3,7 @@ namespace EmsScout.Domain;
 public sealed record AirConditionerCard(
     string Building,
     string SubArea,
-    int? Floor,
+    double? Floor,
     string Page,
     string Name,
     string SwitchState,
